Game mechs transfered from int to float, added time dependent movement and ball movement function

+++ Engine/Vec2.cpp
@@ -1,6 +1,6 @@
 #include "Vec2.h"
 
-Vec2::Vec2(int in_x, int in_y)
+Vec2::Vec2(float in_x, float in_y)
 {
 	x = in_x;
 	y = in_y;
@@ -30,12 +30,12 @@ Vec2 & Vec2::operator-=(const Vec2 & rhs)
 	return *this;
 }
 
-Vec2 Vec2::operator*(const int rhs) const
+Vec2 Vec2::operator*(const float rhs) const
 {
 	return Vec2(x * rhs, y * rhs);
 }
 
-Vec2 Vec2::operator/(const int rhs) const
+Vec2 Vec2::operator/(const float rhs) const
 {
 	return Vec2(x / rhs, y / rhs);
 }
